Added support for ChildActivity::{run,exec} including tests.
Note that we need to disable ASLR for ChildActivity::run since we are
using fork *and* exec and therefore reexecuting the binary. At the same
time, we rely on the function to call having the same address as in the
old process. Since this is a testing-only feature, I think it's fine to
disable ASLR for it.